route changes + sidebar

+++ app/Http/Controllers/Admin/RiderController.php
@@ -346,12 +346,9 @@ class RiderController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit(Rider $rider,Request $request,Rider_detail $rider_detail,$id)
-    {    
-        $rider=Rider::find($id);
-        if(isset($rider)){
-            $rider_detail=$rider->Rider_detail()->get()->first();
-        }        
+    public function edit(Rider $rider,Request $request)
+    {
+        $rider_detail=$rider->Rider_detail;
         $sim_date=$rider->Sim_History()->where('status','active')->get()->first();
         $sim_number=null;
         if(isset($sim_date)){
@@ -578,18 +575,17 @@ class RiderController extends Controller
     {
         $rider_id=$rider->id;
         $client_id=$rider->clients()->get()->first();
-        $client_history=Client_History::where('rider_id', $rider_id)
-        ->where('client_id', $client_id->id)
-        ->where("status","active")
-        ->get()
-        ->first();
-        $client_history->status="deactive";
-        $client_history->deassign_date=Carbon::now()->format("Y-m-d");
-        
-        $client_history->save();
-        
-        $clients = $rider->clients()->detach();
-
+        if(isset($client_id)){
+            $client_history=Client_History::where('rider_id', $rider_id)
+            ->where('client_id', $client_id->id)
+            ->where("status","active")
+            ->get()
+            ->first();
+            $client_history->status="deactive";
+            $client_history->deassign_date=Carbon::now()->format("Y-m-d");
+            $client_history->save();
+            $clients = $rider->clients()->detach();
+        }
         $assign_bike=$rider->Assign_bike()->where('status','active')->get()->first();
         if(isset($assign_bike)){
             $assign_bike->status='deactive';
